sphere

+++ includes/doom.h
@@ -6,7 +6,7 @@
 /*   By: dacuvill <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/12/15 13:08:35 by roduquen          #+#    #+#             */
-/*   Updated: 2019/12/17 02:25:12 by roduquen         ###   ########.fr       */
+/*   Updated: 2019/12/17 11:57:49 by roduquen         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -162,7 +162,7 @@ struct						s_doom
 	t_vec3d					sun;
 	t_light					*sun_light;
 	t_light					*player_light;
-	int						power[6];
+	double					power[6];
 	char					photo;
 	int						actual_i;
 	int						actual_j;
